tickからtick+sequenceidに変更
snapshotがeventと別経路でクライアントに届く関係で同じtickでの順序を守る必要がでた

+++ moorestech_client/Assets/Scripts/Client.Game/InGame/Train/Unit/ITrainUnitHashTickGate.cs
@@ -4,6 +4,6 @@ namespace Client.Game.InGame.Train.Unit
     // Gate that decides whether train simulation can advance by one tick.
     public interface ITrainUnitHashTickGate
     {
-        bool CanAdvanceTick(long currentTick);
+        bool CanAdvanceTick(uint currentTick);
     }
 }
